// src/models/response.ts
export type ResponseCode =
  | "url_not_found"
  | "server_error"
  | "ok"
  | "unauthenticated"
  | "authenticated"
  | "unauthorized"
  | "forbidden"
  | "user_not_found"
  | "already_logged_in"
  | "invalid_req_body"
  | "invalid_credentials"
  | "jwt_error"
  | "created"
  | "updated"
  | "deleted"
  | "duplicate_key"
  | "invalid_req_query"
  | "invalid_req_params"
  | "verification_token_error"
  | "email_already_verified"
  | "email_mismatch"
  | "document_not_found"
  | "no_user_role"
  | "username_already_taken"
  | "email_already_taken"
  | "new_email_not_found"
  | "current_password_incorrect"
  | "passwords_mismatch"
  | "record_not_found"
  | "Database_error";

export interface ResponseObject {
  status: "success" | "fail" | "error";
  code: ResponseCode;
  message: string;
  [key: string]: any;
}